Refactor to use custom hook instead
