+++ packages/web/hooks/index.js
@@ -1,3 +1,4 @@
 export { default as useTheme } from './useTheme';
 export { default as useModal } from './useModal';
 export { default as useAuth } from './useAuth';
+export { default as useVisibleComponent } from './useVisibleComponent';
